refactor(credentials): share the parts every credential form repeats
The OAuth2 modal arrived as a near-copy of the Entra one, which put new
duplication at 22.8% against a 3% gate. The duplicated parts are the ones
that have nothing to do with which grant is being configured: the test
outcome line, the used-by panel, the header delete button and the submit
label.

Moved to `credential-modal-parts.tsx` and used from both. The per-type
fields stay in their own modals, which is the half that genuinely differs
— so the next credential type costs its fields rather than another copy of
the shell.
